+++ sdk/src/hooks/useOrder.ts
@@ -11,9 +11,8 @@ import {
 import { inboxABI } from '../constants/abis.js'
 import { typeHash } from '../constants/typehash.js'
 import { useOmniContext } from '../context/omni.js'
-import type { Order } from '../types/order.js'
-import type { OrderStatus } from '../types/order.js'
-import type { OptionalAbis } from '../types/order.js'
+import type { OptionalAbis } from '../types/abi.js'
+import type { Order, OrderStatus } from '../types/order.js'
 import { encodeOrder } from '../utils/encodeOrder.js'
 import { useDidFill } from './useDidFill.js'
 import { useGetOpenOrder } from './useGetOpenOrder.js'
