+++ sdk/src/hooks/useValidateOrder.ts
@@ -3,7 +3,8 @@ import { useMemo } from 'react'
 import { encodeFunctionData, zeroAddress } from 'viem'
 import { useOmniContext } from '../context/omni.js'
 import { type FetchJSONError, fetchJSON } from '../internal/api.js'
-import type { OptionalAbis, Order } from '../types/order.js'
+import type { OptionalAbis } from '../types/abi.js'
+import type { Order } from '../types/order.js'
 import { isContractCall } from '../types/order.js'
 import { toJSON } from './util.js'
